move SpawnNSAndNetwork to a new module
